Refactor and enhance GraphQL sync functionality
- Renamed test function to better reflect its purpose regarding an empty repository catalog.
- Added a new test to validate the resolution of a registry-backed repository when no catalog row exists.
- Modularized GraphQL operations by introducing new files for client interactions, documents, progress handling, and subscriptions.
- Implemented a `SyncProgressRenderer` for improved interactive sync progress display.
- Enhanced error handling and logging for GraphQL operations, ensuring better feedback during sync tasks.
- Added comprehensive tests to validate the new GraphQL functionality and ensure accurate reporting of sync progress.
