+++ bitloops/src/cli/devql/graphql/documents.rs
@@ -1,0 +1,246 @@
+pub(super) const INIT_SCHEMA_MUTATION: &str = r#"
+    mutation InitSchema {
+      initSchema {
+        success
+        repoIdentity
+        repoId
+        relationalBackend
+        eventsBackend
+      }
+    }
+"#;
+
+pub(super) const INGEST_MUTATION: &str = r#"
+    mutation Ingest($input: IngestInput!) {
+      ingest(input: $input) {
+        success
+        checkpointsProcessed
+        eventsInserted
+        artefactsUpserted
+        checkpointsWithoutCommit
+        temporaryRowsPromoted
+        semanticFeatureRowsUpserted
+        semanticFeatureRowsSkipped
+        symbolEmbeddingRowsUpserted
+        symbolEmbeddingRowsSkipped
+        symbolCloneEdgesUpserted
+        symbolCloneSourcesScored
+      }
+    }
+"#;
+
+pub(super) const ENQUEUE_SYNC_MUTATION: &str = r#"
+    mutation EnqueueSync($input: EnqueueSyncInput!) {
+      enqueueSync(input: $input) {
+        merged
+        task {
+          taskId
+          repoId
+          repoName
+          repoIdentity
+          source
+          mode
+          status
+          phase
+          submittedAtUnix
+          startedAtUnix
+          updatedAtUnix
+          completedAtUnix
+          queuePosition
+          tasksAhead
+          currentPath
+          pathsTotal
+          pathsCompleted
+          pathsRemaining
+          pathsUnchanged
+          pathsAdded
+          pathsChanged
+          pathsRemoved
+          cacheHits
+          cacheMisses
+          parseErrors
+          error
+          summary {
+            success
+            mode
+            parserVersion
+            extractorVersion
+            activeBranch
+            headCommitSha
+            headTreeSha
+            pathsUnchanged
+            pathsAdded
+            pathsChanged
+            pathsRemoved
+            cacheHits
+            cacheMisses
+            parseErrors
+            validation {
+              valid
+              expectedArtefacts
+              actualArtefacts
+              expectedEdges
+              actualEdges
+              missingArtefacts
+              staleArtefacts
+              mismatchedArtefacts
+              missingEdges
+              staleEdges
+              mismatchedEdges
+              filesWithDrift {
+                path
+                missingArtefacts
+                staleArtefacts
+                mismatchedArtefacts
+                missingEdges
+                staleEdges
+                mismatchedEdges
+              }
+            }
+          }
+        }
+      }
+    }
+"#;
+
+pub(super) const SYNC_TASK_QUERY: &str = r#"
+    query SyncTask($id: String!) {
+      syncTask(id: $id) {
+        taskId
+        repoId
+        repoName
+        repoIdentity
+        source
+        mode
+        status
+        phase
+        submittedAtUnix
+        startedAtUnix
+        updatedAtUnix
+        completedAtUnix
+        queuePosition
+        tasksAhead
+        currentPath
+        pathsTotal
+        pathsCompleted
+        pathsRemaining
+        pathsUnchanged
+        pathsAdded
+        pathsChanged
+        pathsRemoved
+        cacheHits
+        cacheMisses
+        parseErrors
+        error
+        summary {
+          success
+          mode
+          parserVersion
+          extractorVersion
+          activeBranch
+          headCommitSha
+          headTreeSha
+          pathsUnchanged
+          pathsAdded
+          pathsChanged
+          pathsRemoved
+          cacheHits
+          cacheMisses
+          parseErrors
+          validation {
+            valid
+            expectedArtefacts
+            actualArtefacts
+            expectedEdges
+            actualEdges
+            missingArtefacts
+            staleArtefacts
+            mismatchedArtefacts
+            missingEdges
+            staleEdges
+            mismatchedEdges
+            filesWithDrift {
+              path
+              missingArtefacts
+              staleArtefacts
+              mismatchedArtefacts
+              missingEdges
+              staleEdges
+              mismatchedEdges
+            }
+          }
+        }
+      }
+    }
+"#;
+
+pub(super) const SYNC_PROGRESS_SUBSCRIPTION: &str = r#"
+    subscription SyncProgress($taskId: String!) {
+      syncProgress(taskId: $taskId) {
+        taskId
+        repoId
+        repoName
+        repoIdentity
+        source
+        mode
+        status
+        phase
+        submittedAtUnix
+        startedAtUnix
+        updatedAtUnix
+        completedAtUnix
+        queuePosition
+        tasksAhead
+        currentPath
+        pathsTotal
+        pathsCompleted
+        pathsRemaining
+        pathsUnchanged
+        pathsAdded
+        pathsChanged
+        pathsRemoved
+        cacheHits
+        cacheMisses
+        parseErrors
+        error
+        summary {
+          success
+          mode
+          parserVersion
+          extractorVersion
+          activeBranch
+          headCommitSha
+          headTreeSha
+          pathsUnchanged
+          pathsAdded
+          pathsChanged
+          pathsRemoved
+          cacheHits
+          cacheMisses
+          parseErrors
+          validation {
+            valid
+            expectedArtefacts
+            actualArtefacts
+            expectedEdges
+            actualEdges
+            missingArtefacts
+            staleArtefacts
+            mismatchedArtefacts
+            missingEdges
+            staleEdges
+            mismatchedEdges
+            filesWithDrift {
+              path
+              missingArtefacts
+              staleArtefacts
+              mismatchedArtefacts
+              missingEdges
+              staleEdges
+              mismatchedEdges
+            }
+          }
+        }
+      }
+    }
+"#;
