+++ bitloops/src/host/devql/db_utils.rs
@@ -296,8 +296,7 @@ pub(super) async fn sqlite_exec_path_inner(
             )
         }
         .with_context(|| format!("opening SQLite database at {}", db_path.display()))?;
-        conn.busy_timeout(std::time::Duration::from_secs(30))
-            .context("setting SQLite busy timeout")?;
+        configure_sqlite_connection(&conn, "executing SQLite statements")?;
         conn.execute_batch(&statement)
             .context("executing SQLite statements")?;
         Ok(())
@@ -325,8 +324,7 @@ pub(super) async fn sqlite_exec_batch_transactional_path(
             rusqlite::OpenFlags::SQLITE_OPEN_READ_WRITE,
         )
         .with_context(|| format!("opening SQLite database at {}", db_path.display()))?;
-        conn.busy_timeout(std::time::Duration::from_secs(30))
-            .context("setting SQLite busy timeout")?;
+        configure_sqlite_connection(&conn, "executing SQLite transactional batch")?;
         conn.execute_batch("PRAGMA journal_mode = WAL; PRAGMA synchronous = NORMAL;")
             .context("configuring SQLite WAL transaction settings")?;
 
@@ -405,8 +403,7 @@ pub(crate) async fn sqlite_query_rows_path(path: &Path, sql: &str) -> Result<Vec
             rusqlite::OpenFlags::SQLITE_OPEN_READ_WRITE,
         )
             .with_context(|| format!("opening SQLite database at {}", db_path.display()))?;
-        conn.busy_timeout(std::time::Duration::from_secs(30))
-            .context("setting SQLite busy timeout")?;
+        configure_sqlite_connection(&conn, "querying SQLite rows")?;
         let mut stmt = conn.prepare(&query).context("preparing SQLite query")?;
         let column_names = stmt
             .column_names()
@@ -431,6 +428,14 @@ pub(crate) async fn sqlite_query_rows_path(path: &Path, sql: &str) -> Result<Vec
     })
     .await
     .context("joining SQLite query task")?
+}
+
+fn configure_sqlite_connection(conn: &rusqlite::Connection, context: &str) -> Result<()> {
+    conn.busy_timeout(std::time::Duration::from_secs(30))
+        .with_context(|| format!("setting SQLite busy timeout while {context}"))?;
+    conn.execute_batch("PRAGMA foreign_keys = ON;")
+        .with_context(|| format!("enabling SQLite foreign keys while {context}"))?;
+    Ok(())
 }
 
 pub(crate) fn sqlite_value_to_json(value: rusqlite::types::ValueRef<'_>) -> Value {
