+++ bitloops/src/host/devql/ingestion/schema/relational_postgres_schema.rs
@@ -254,7 +254,8 @@ CREATE TABLE IF NOT EXISTS repo_sync_state (
     last_sync_started_at TEXT,
     last_sync_completed_at TEXT,
     last_sync_status TEXT,
-    last_sync_reason TEXT
+    last_sync_reason TEXT,
+    FOREIGN KEY (repo_id) REFERENCES repositories(repo_id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS content_cache (
