+++ bitloops/src/host/devql/sync/schema.rs
@@ -11,7 +11,8 @@ CREATE TABLE IF NOT EXISTS repo_sync_state (
     last_sync_started_at TEXT,
     last_sync_completed_at TEXT,
     last_sync_status TEXT,
-    last_sync_reason TEXT
+    last_sync_reason TEXT,
+    FOREIGN KEY (repo_id) REFERENCES repositories(repo_id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS current_file_state (
@@ -29,7 +30,8 @@ CREATE TABLE IF NOT EXISTS current_file_state (
     exists_in_index INTEGER NOT NULL,
     exists_in_worktree INTEGER NOT NULL,
     last_synced_at TEXT NOT NULL,
-    PRIMARY KEY (repo_id, path)
+    PRIMARY KEY (repo_id, path),
+    FOREIGN KEY (repo_id) REFERENCES repositories(repo_id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS content_cache (
@@ -83,6 +85,27 @@ CREATE TABLE IF NOT EXISTS content_cache_edges (
 "#
 }
 
+pub(crate) fn sync_repo_sync_state_migration_sql() -> &'static str {
+    r#"
+DROP TABLE IF EXISTS repo_sync_state;
+
+CREATE TABLE IF NOT EXISTS repo_sync_state (
+    repo_id TEXT PRIMARY KEY,
+    repo_root TEXT NOT NULL,
+    active_branch TEXT,
+    head_commit_sha TEXT,
+    head_tree_sha TEXT,
+    parser_version TEXT NOT NULL,
+    extractor_version TEXT NOT NULL,
+    last_sync_started_at TEXT,
+    last_sync_completed_at TEXT,
+    last_sync_status TEXT,
+    last_sync_reason TEXT,
+    FOREIGN KEY (repo_id) REFERENCES repositories(repo_id) ON DELETE CASCADE
+);
+"#
+}
+
 pub(crate) fn sync_current_file_state_migration_sql() -> &'static str {
     r#"
 DROP TABLE IF EXISTS current_file_state;
@@ -102,7 +125,8 @@ CREATE TABLE IF NOT EXISTS current_file_state (
     exists_in_index INTEGER NOT NULL,
     exists_in_worktree INTEGER NOT NULL,
     last_synced_at TEXT NOT NULL,
-    PRIMARY KEY (repo_id, path)
+    PRIMARY KEY (repo_id, path),
+    FOREIGN KEY (repo_id) REFERENCES repositories(repo_id) ON DELETE CASCADE
 );
 "#
 }
@@ -132,7 +156,8 @@ CREATE TABLE IF NOT EXISTS artefacts_current (
     docstring TEXT,
     updated_at TEXT NOT NULL,
     PRIMARY KEY (repo_id, path, symbol_id),
-    UNIQUE (repo_id, artefact_id)
+    UNIQUE (repo_id, artefact_id),
+    FOREIGN KEY (repo_id) REFERENCES repositories(repo_id) ON DELETE CASCADE
 );
 
 CREATE INDEX IF NOT EXISTS artefacts_current_path_idx
@@ -167,7 +192,8 @@ CREATE TABLE IF NOT EXISTS artefact_edges_current (
         (start_line IS NULL AND end_line IS NULL)
         OR (start_line IS NOT NULL AND end_line IS NOT NULL AND start_line > 0 AND end_line >= start_line)
     ),
-    PRIMARY KEY (repo_id, edge_id)
+    PRIMARY KEY (repo_id, edge_id),
+    FOREIGN KEY (repo_id) REFERENCES repositories(repo_id) ON DELETE CASCADE
 );
 
 CREATE INDEX IF NOT EXISTS artefact_edges_current_path_idx
